# Saving a bit more memory in classical calculations